requirements.txt
aiogram==2.5
openai==1.55.0
python-dotenv